First attempt at the load_gadm function.  Not yet working for ISO_A2_L2 and above (which is weird).  Also seems to only be pulling a subset of what I expect.
@shauntruelove  I made a few small modifications to the
standardzie_name and create_standardized_name functions.  The main
thing I did was add a call to database_lookup_location_name, which
just checks the database to see if the object in question is there.
For now it's pretty fast, and avoids some slow down, but I could see
that changing in the future.

Was going to commit the database, but even with only a few
countries/ISO_A2_L1 regions, the database is too big to fit on
github.  I guess we'll need to settle for writing functions to pull
the important data.